3.29 update lvpinhe

+++ travel-common/travel-common-dao/src/main/java/com/zb/mapper/HotelMapper.java
@@ -8,7 +8,7 @@ import org.apache.ibatis.annotations.Mapper;
 @Mapper
 public interface HotelMapper {
 
-	public Hotel getHotelById(@Param(value = "id") Long id)throws Exception;
+	public Hotel getHotelById(@Param(value = "id") Integer id)throws Exception;
 
 	public List<Hotel>	getHotelListByMap(Map<String,Object> param)throws Exception;
